+++ src/main/webapp/resources/css/pages/seller/mypage.css
@@ -1,48 +1,39 @@
 @charset "UTF-8";
 
-form {
-	width: 780px;
+.mypage-container {
 	padding: 40px;
+	width: 780px;
 }
 
-h3 {
-	font-weight: bold;
-	font-size: 32px;
+h2 {
 	padding-left: 20px;
-	margin-bottom: 0px;
+	font-size: 25px;
 }
 
-h2, p {
-	padding-left: 20px;
-	margin-top: 0px;
+.user-info {
+	padding: 20px 0px 15px 20px;
 }
 
-.menu-label {
+.user-info label {
+	color: #ff9670;
 	font-weight: bold;
 	font-size: 25px;
-	margin-bottom: 10px;
-	padding: 20px 0px 10px 20px;
-}
-
-label {
-	font-weight: bold;
-	font-size: 25px;
-	margin-bottom: 10px;
-	padding: 20px 0px 20px 0px;
-}
-
-input[type="text"] {
-	width: 100%;
-	padding: 10px;
-	font-size: 20px;
-	margin-top: 5px;
-	margin-bottom: 10px;
-	border: 1px solid #ccc;
-	border-radius: 6px;
 }
 
 .input-group {
-	padding: 0px 0px 20px 20px;
+	display: block;
+	width: 100%;
+	margin-top: 10px;
+}
+
+.input-group input {
+	flex: 1;
+	padding: 10px;
+	font-size: 20px;
+	border: 1px solid #ccc;
+	border-radius: 6px;
+	width: 100%;
+	box-sizing: border-box;
 }
 
 .btn-group {
@@ -60,7 +51,7 @@ input[type="text"] {
 	width: 200px;
 	border-radius: 6px;
 	cursor: pointer;
-	font-size: 20px;
+	font-size: 30px;
 	font-weight: bold;
 }
 
@@ -76,6 +67,30 @@ input[type="text"] {
 
 .btn.filled:hover {
 	background-color: #ff7a52;
+	color: white;
+}
+
+.input-group.with-btn {
+	display: flex;
+	align-items: center;
+	gap: 10px;
+}
+
+/* 버튼이 오른쪽에 붙는 경우 */
+.btn.side-btn {
+	width: auto;
+	padding: 10px 16px;
+	font-size: 18px;
+	border: 2px solid #ff9670;
+	background: transparent;
+	color: #ff9670;
+	border-radius: 6px;
+	cursor: pointer;
+	font-weight: bold;
+}
+
+.btn.side-btn:hover {
+	background-color: #ff9670;
 	color: white;
 }
